requirements.txt
spotipy
pyyaml
python-frontmatter
python-dotenv
